서버 움직임 구현
체스말의 움직임을 서버에서 관리하도록 구현.
현재 하나의 패킷으로 데이터를 송수신하는데, 패킷이 여러개가 될 경우 어떻게 처리할지 고려해야하는 상황

+++ ChessGame/Client/Renderer.cpp
@@ -1,11 +1,12 @@
 #include "stdafx.h"
 #include "Renderer.h"
+#include <cstdlib>
+#include <fstream>
 
 CRenderer::CRenderer(int windowSizeX, int windowSizeY)
 {
 	Initialize(windowSizeX, windowSizeY);
 }
-
 
 CRenderer::~CRenderer()
 {
@@ -60,7 +61,7 @@ void CRenderer::AddShader(GLuint ShaderProgram, const char* pShaderText, GLenum 
 	const GLchar* p[1];
 	p[0] = pShaderText;
 	GLint Lengths[1];
-	Lengths[0] = strlen(pShaderText);
+	Lengths[0] = (GLint)strlen(pShaderText);
 	//쉐이더 코드를 쉐이더 오브젝트에 할당
 	glShaderSource(ShaderObj, 1, p, Lengths);
